Add log nesting, ports can use custom drivers
