+++ revvy/robot/configurations.py
@@ -1,9 +1,12 @@
 # SPDX-License-Identifier: GPL-3.0-only
+from revvy.robot.ports.motors.dc_motor import DcMotorController
+from revvy.robot.ports.sensors.ev3 import ev3_color, Ev3UARTSensor
+from revvy.robot.ports.sensors.simple import hcsr04, bumper_switch
 
 
 class Motors:
     RevvyMotor = {
-        'driver': 'DcMotor',
+        'driver': DcMotorController,
         'config': {
             'speed_controller': [0.6065, 0.3935, 0, -150, 150],
             'position_controller': [0.1, 0.0000, 0, -150, 150],
@@ -15,7 +18,7 @@ class Motors:
         }
     }
     RevvyMotor_CCW = {
-        'driver': 'DcMotor',
+        'driver': DcMotorController,
         'config': {
             'speed_controller': [0.6065, 0.3935, 0, -150, 150],
             'position_controller': [0.1, 0.0000, 0, -150, 150],
@@ -29,7 +32,7 @@ class Motors:
 
 
 class Sensors:
-    HC_SR04 = {'driver': 'HC_SR04', 'config': {}}
-    BumperSwitch = {'driver': 'BumperSwitch', 'config': {}}
-    EV3 = {'driver': 'EV3', 'config': {}}
-    EV3_Color = {'driver': 'EV3_Color', 'config': {}}
+    HC_SR04 = {'driver': hcsr04, 'config': {}}
+    BumperSwitch = {'driver': bumper_switch, 'config': {}}
+    EV3 = {'driver': lambda port, _: Ev3UARTSensor(port), 'config': {}}
+    EV3_Color = {'driver': ev3_color, 'config': {}}
