+++ revvy/robot/ports/sensors/base.py
@@ -4,8 +4,8 @@ from revvy.robot.ports.common import PortDriver, PortInstance
 
 
 class NullSensor(PortDriver):
-    def __init__(self):
-        super().__init__('NotConfigured')
+    def __init__(self, port):
+        super().__init__(port, 'NotConfigured')
 
     def on_port_type_set(self):
         pass
@@ -24,15 +24,11 @@ class NullSensor(PortDriver):
 
 class BaseSensorPortDriver(PortDriver):
     def __init__(self, driver, port: PortInstance):
-        super().__init__(driver)
-        self._log_tag = f'[driver={driver}]: '
+        super().__init__(port, driver)
         self._port = port
         self._interface = port.interface
         self._value = None
         self._raw_value = None
-
-    def log(self, message):
-        self._port.log(self._log_tag + message)
 
     def on_port_type_set(self):
         pass
